refactor(async_rl_loop): inline AsyncRolloutScheduler, delete datum.py and rewards.py
Make async_rl_loop.py fully self-contained for RL logic:

- Inline AsyncRolloutScheduler + stream_groups (async-only, was in rollout.py)
- Delete datum.py (build_prompt_group inlined in previous commit)
- Delete rewards.py (default reward/filter inlined in previous commit)
- Config.reward_fn and Config.filter_fn now default to None — each loop
  file provides its own defaults, making the RL logic visible to users
- rollout.py kept minimal: collect_sync_batch + RolloutStats only (used
  by sync path in train.py)

Design philosophy: users read ONE file (rl_loop or async_rl_loop) and
understand the full RL flow. Utils hide plumbing, not RL logic.
