+++ training/recipes/async_rl_loop.py
@@ -71,7 +71,6 @@ from training.utils.rl.config import Config
 from training.utils.rl.losses import PromptGroup, build_loss_fn, resolve_builtin_loss
 from training.utils.rl.router_replay import build_r3_routing_matrices
 from training.utils.rl.train import TrainContext, train_one_group, finish_step
-from training.utils.rl.rollout import AsyncRolloutScheduler
 
 logger = logging.getLogger(__name__)
 
@@ -229,6 +228,209 @@ def build_prompt_group(
         completions=[s.text for s in sampled] if trajectory_dir else None,
         row_meta={"ground_truth": row.get("ground_truth", "")} if trajectory_dir else None,
     )
+
+
+# ---------------------------------------------------------------------------
+# Async rollout scheduler (AReaL-style capacity gating)
+#
+# Manages in-flight sampling tasks with two-level gating:
+#   1. Staleness cap (policy window): how far ahead rollouts can be
+#   2. Concurrency cap (resource window): max in-flight HTTP requests
+#
+# stream_groups() yields accepted groups one at a time for immediate
+# training — the core of the streaming pipeline.
+# ---------------------------------------------------------------------------
+
+
+class AsyncRolloutScheduler:
+    """Async rollout scheduler with AReaL-style capacity gating."""
+
+    def __init__(
+        self,
+        step_target: int,
+        max_head_offpolicy_versions: int,
+        filter_fn: Callable | None = None,
+        global_step: int = 0,
+        total_accepted: int = 0,
+        total_rejected: int = 0,
+        rows_submitted: int = 0,
+        max_concurrent: int | None = None,
+    ):
+        self._step_target = step_target
+        self._max_offpolicy = max_head_offpolicy_versions
+        self._filter_fn = filter_fn
+
+        self._current_version = global_step
+        self._total_accepted = total_accepted
+        self._total_rejected = total_rejected
+        self._rows_submitted = rows_submitted
+
+        policy_window = (max_head_offpolicy_versions + 1) * step_target
+        self._max_concurrent = min(max_concurrent, policy_window) if max_concurrent is not None else policy_window
+
+        self._in_flight: set[asyncio.Task] = set()
+        self._result_queue: asyncio.Queue[tuple[PromptGroup | None, int]] = asyncio.Queue()
+        self._rows_exhausted = False
+
+    @property
+    def current_version(self) -> int:
+        return self._current_version
+
+    @property
+    def data_exhausted(self) -> bool:
+        return self._rows_exhausted and len(self._in_flight) == 0
+
+    def _staleness_cap(self) -> int:
+        budget = (
+            (self._max_offpolicy + self._current_version + 1) * self._step_target
+            - (self._total_accepted + len(self._in_flight))
+        )
+        return max(budget, 0)
+
+    def _concurrency_cap(self) -> int:
+        return max(self._max_concurrent - len(self._in_flight), 0)
+
+    def _capacity(self) -> int:
+        return min(self._staleness_cap(), self._concurrency_cap())
+
+    def _submit_one(self, sample_fn_factory, row: dict) -> None:
+        version = self._current_version
+        coro = sample_fn_factory(row)
+
+        async def _worker():
+            try:
+                result = await coro
+            except Exception as exc:
+                logger.warning("Rollout task failed (%s): %s", type(exc).__name__, exc or repr(exc))
+                result = None
+            self._result_queue.put_nowait((result, version))
+
+        task = asyncio.create_task(_worker())
+        self._in_flight.add(task)
+        task.add_done_callback(self._in_flight.discard)
+        self._rows_submitted += 1
+
+    async def stream_groups(self, sample_fn_factory, rows):
+        """Yield accepted (PromptGroup, version) one at a time.
+
+        The core of the streaming pipeline: each accepted group is yielded
+        immediately so the caller can send it to the trainer for fwd_bwd
+        while sampling continues in the background.
+        """
+        accepted = 0
+
+        def _try_submit():
+            if self._rows_exhausted:
+                return
+            cap = self._capacity()
+            for _ in range(cap):
+                try:
+                    row = next(rows)
+                except StopIteration:
+                    self._rows_exhausted = True
+                    break
+                self._submit_one(sample_fn_factory, row)
+
+        _try_submit()
+
+        while accepted < self._step_target:
+            if not self._in_flight and self._result_queue.empty():
+                break
+
+            try:
+                item, version = await asyncio.wait_for(
+                    self._result_queue.get(), timeout=0.1,
+                )
+            except asyncio.TimeoutError:
+                _try_submit()
+                continue
+
+            if item is None:
+                _try_submit()
+                continue
+
+            if self._filter_fn is not None and not self._filter_fn(item):
+                self._total_rejected += 1
+                _try_submit()
+                continue
+
+            self._total_accepted += 1
+            accepted += 1
+            yield item, version
+            _try_submit()
+
+    async def collect_batch(self, sample_fn_factory, rows):
+        """Collect step_target accepted groups (blocking).
+
+        Re-implements the batch collection with full stats tracking.
+        """
+        from training.utils.rl.rollout import RolloutStats
+
+        accepted = []
+        stats = RolloutStats()
+        t0 = _time.time()
+
+        def _process_one(item, version):
+            if item is None:
+                stats.sample_fails += 1
+                stats.total_sampled += 1
+                return
+            stats.total_sampled += 1
+            stats.raw_rewards.extend(item.rewards)
+            if self._filter_fn is not None and not self._filter_fn(item):
+                stats.filter_drops += 1
+                self._total_rejected += 1
+                return
+            accepted.append(item)
+            self._total_accepted += 1
+            stats.version_offsets.append(self._current_version - version)
+
+        def _drain():
+            while not self._result_queue.empty():
+                pair = self._result_queue.get_nowait()
+                _process_one(*pair)
+                if len(accepted) >= self._step_target:
+                    break
+
+        _drain()
+
+        while len(accepted) < self._step_target:
+            if not self._rows_exhausted:
+                cap = self._capacity()
+                for _ in range(cap):
+                    try:
+                        row = next(rows)
+                    except StopIteration:
+                        self._rows_exhausted = True
+                        break
+                    self._submit_one(sample_fn_factory, row)
+
+            if not self._in_flight and self._result_queue.empty():
+                break
+
+            try:
+                item, version = await asyncio.wait_for(self._result_queue.get(), timeout=0.1)
+                _process_one(item, version)
+                if len(accepted) < self._step_target:
+                    _drain()
+            except asyncio.TimeoutError:
+                _drain()
+
+        stats.valid_groups = len(accepted)
+        stats.wall_time = _time.time() - t0
+        return accepted, stats
+
+    def bump_version(self) -> None:
+        """Increment version after optim_step. Opens capacity budget."""
+        self._current_version += 1
+
+    def get_state(self) -> dict:
+        """Snapshot counters for checkpointing."""
+        return {
+            "rows_submitted": self._rows_submitted,
+            "total_accepted": self._total_accepted,
+            "total_rejected": self._total_rejected,
+        }
 
 
 # ---------------------------------------------------------------------------
