+++ training/utils/rl/__init__.py
@@ -23,13 +23,7 @@ __all__ = [
     "train_one_step",
     "run_rl_loop",
     # Rollout scheduling
-    "AsyncRolloutScheduler",
     "RolloutStats",
-    # Datum building
-    "build_prompt_group",
-    # Rewards
-    "default_math_reward",
-    "default_variance_filter",
     # Metrics helpers
     "add_response_length_stats",
     "add_train_perf_metrics",
@@ -50,13 +44,8 @@ from training.utils.rl.train import (
     train_one_step,
     run_rl_loop,
 )
-from training.utils.rl.rollout import (
-    AsyncRolloutScheduler,
-    RolloutStats,
-)
+from training.utils.rl.rollout import RolloutStats
 from training.utils.rl.losses import PromptGroup
-from training.utils.rl.datum import build_prompt_group
-from training.utils.rl.rewards import default_math_reward, default_variance_filter
 from training.utils.rl.metrics import (
     build_loop_metrics,
     total_target_tokens,
